pyproject.toml
[project]
name = "ap-invoice-agent"
version = "0.1.0"
description = "Agentic AI Accounts Payable Invoice Processing with MCP"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.2.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
    "Pillow>=11.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "api", "mcp_erp_server"]
